+++ packages/core/src/actions/smartAccount/sendTransaction.ts
@@ -10,6 +10,7 @@ import { isBaseSmartAccountClient } from "../../client/isSmartAccountClient.js";
 import { AccountNotFoundError } from "../../errors/account.js";
 import { IncompatibleClientError } from "../../errors/client.js";
 import { TransactionMissingToParamError } from "../../errors/transaction.js";
+import type { UserOperationOverrides } from "../../types.js";
 import { buildUserOperationFromTx } from "./buildUserOperationFromTx.js";
 import { _sendUserOperation } from "./internal/sendUserOperation.js";
 import { waitForUserOperationTransaction } from "./waitForUserOperationTransacation.js";
@@ -22,8 +23,9 @@ export const sendTransaction: <
   TChainOverride extends Chain | undefined = Chain | undefined
 >(
   client: Client<Transport, TChain, TAccount>,
-  args: SendTransactionParameters<TChain, TAccount, TChainOverride>
-) => Promise<Hex> = async (client, args) => {
+  args: SendTransactionParameters<TChain, TAccount, TChainOverride>,
+  overrides?: UserOperationOverrides
+) => Promise<Hex> = async (client, args, overrides) => {
   const { account = client.account } = args;
   if (!account || typeof account === "string") {
     throw new AccountNotFoundError();
@@ -41,7 +43,7 @@ export const sendTransaction: <
     );
   }
 
-  const uoStruct = await buildUserOperationFromTx(client, args);
+  const uoStruct = await buildUserOperationFromTx(client, args, overrides);
   const { hash } = await _sendUserOperation(client, {
     account: account as SmartContractAccount,
     uoStruct,