+++ site/snippets/signers/alchemy/SignupLoginComponent.tsx
@@ -1,37 +1,47 @@
 import { AlchemySigner } from "@alchemy/aa-alchemy";
 import { useMutation } from "@tanstack/react-query";
-import { useState } from "react";
-import { usePromise } from "./usePromise.js";
+import { useEffect, useMemo, useState } from "react";
 
 export const SignupLoginComponent = () => {
-  // The usePromise hook is a helpful utility that makes it easier to resolve user's input after
-  // a request has already been initiated
-  const { promise: bundle, resolve } = usePromise<string>();
   const [email, setEmail] = useState<string>("");
-  const [bundleInput, setBundleInput] = useState<string>("");
 
   // It is recommended you wrap this in React Context or other state management
-  const signer = new AlchemySigner({
-    client: {
-      connection: {
-        jwt: "alcht_<KEY>",
-      },
-      iframeConfig: {
-        iframeContainerId: "turnkey-iframe-container",
-      },
-    },
-  });
+  const signer = useMemo(
+    () =>
+      new AlchemySigner({
+        client: {
+          connection: {
+            jwt: "alcht_<KEY>",
+          },
+          iframeConfig: {
+            iframeContainerId: "turnkey-iframe-container",
+          },
+        },
+      }),
+    []
+  );
 
   // we are using react-query to handle loading states more easily, but feel free to use w/e state management library you prefer
   const { mutate: loginOrSignup, isLoading } = useMutation({
     mutationFn: (email: string) =>
-      signer.authenticate({ type: "email", email, bundle }),
+      signer.authenticate({ type: "email", email }),
   });
 
-  // The below view allows you to collect the email from the user and then the OTP bundle that they are emailed
+  useEffect(() => {
+    const urlParams = new URLSearchParams(window.location.search);
+    if (urlParams.has("bundle")) {
+      // this will complete email auth
+      signer
+        .authenticate({ type: "email", bundle: urlParams.get("bundle")! })
+        // redirect the user or do w/e you want once the user is authenticated
+        .then(() => (window.location.href = "/"));
+    }
+  }, [signer]);
+
+  // The below view allows you to collect the email from the user
   return (
     <>
-      {!isLoading ? (
+      {!isLoading && (
         <div>
           <input
             type="email"
@@ -39,15 +49,6 @@ export const SignupLoginComponent = () => {
             onChange={(e) => setEmail(e.target.value)}
           />
           <button onClick={() => loginOrSignup(email)}>Submit</button>
-        </div>
-      ) : (
-        <div>
-          <input
-            type="text"
-            value={bundleInput}
-            onChange={(e) => setBundleInput(e.target.value)}
-          />
-          <button onClick={() => resolve(bundleInput)}>Submit</button>
         </div>
       )}
       <div id="turnkey-iframe-container" />